Datasources: share diagnostics drawer content
