+++ public/app/features/dashboard-scene/sharing/DownloadDiagnostics.tsx
@@ -1,24 +1,19 @@
-import { css } from '@emotion/css';
-import { useEffect, useRef, useState } from 'react';
+import { useEffect, useRef } from 'react';
 import { useAsyncFn } from 'react-use';
 
-import { type GrafanaTheme2 } from '@grafana/data';
 import { Trans, t } from '@grafana/i18n';
-import { isFetchError } from '@grafana/runtime';
 import {
   sceneGraph,
   type SceneComponentProps,
-  SceneDataTransformer,
-  type SceneObject,
   SceneObjectBase,
   type SceneObjectRef,
-  SceneQueryRunner,
   type VizPanel,
 } from '@grafana/scenes';
 import { type DataQuery } from '@grafana/schema';
-import { Alert, Button, Checkbox, useStyles2 } from '@grafana/ui';
 import { downloadDiagnosticsForQueries } from 'app/features/query/diagnostics/downloadDiagnostics';
 
+import { DiagnosticsDrawerContent } from './DiagnosticsDrawerContent';
+import { getQueryRunnerFor } from './diagnosticsUtils';
 import { type SceneShareTabState, type ShareView } from './types';
 
 export interface DownloadDiagnosticsState extends SceneShareTabState {
@@ -41,69 +36,43 @@ export class DownloadDiagnostics extends SceneObjectBase<DownloadDiagnosticsStat
   }
 }
 
-// Inlined rather than imported from dashboard-scene/utils/utils: that module transitively reaches
-// DashboardScene, which imports ShareDrawer (which imports this view), creating an import cycle.
-function getQueryRunnerFor(sceneObject: SceneObject | undefined): SceneQueryRunner | undefined {
-  if (!sceneObject) {
-    return undefined;
-  }
-  const dataProvider = sceneObject.state.$data ?? sceneObject.parent?.state.$data;
-  if (dataProvider instanceof SceneQueryRunner) {
-    return dataProvider;
-  }
-  if (dataProvider instanceof SceneDataTransformer) {
-    return getQueryRunnerFor(dataProvider);
-  }
-  return undefined;
-}
-
-// The download uses a blob-response fetch, whose FetchError carries the detail in status/statusText
-// (its data is a Blob and message is unset), so build a message from those rather than error.message
-// which would leave the alert body empty.
-function diagnosticsErrorMessage(error: Error): string {
-  if (isFetchError(error)) {
-    const parts = [error.status, error.statusText].filter(Boolean);
-    return parts.length ? parts.join(' ') : t('dashboard.diagnostics.request-failed', 'Request failed');
-  }
-  return error.message || t('dashboard.diagnostics.error-title', 'Failed to generate diagnostics');
-}
-
 function DownloadDiagnosticsRenderer({ model }: SceneComponentProps<DownloadDiagnostics>) {
   const { onDismiss, panelRef } = model.useState();
-  const styles = useStyles2(getStyles);
   const abortRef = useRef<AbortController | null>(null);
-  const [includeLogs, setIncludeLogs] = useState(true);
 
   // Abort any in-flight request if the drawer unmounts.
   useEffect(() => () => abortRef.current?.abort(), []);
 
-  const [{ loading: isGenerating, error }, onDownload] = useAsyncFn(async () => {
-    const panel = panelRef?.resolve();
-    if (!panel) {
-      return;
-    }
-    const runner = getQueryRunnerFor(panel);
-    // Classic panels keep the datasource on the query runner rather than on each target, and unlike
-    // the normal /api/ds/query path nothing fills that in here. Copy the runner-level datasource
-    // onto any query that lacks one so the diagnostics endpoint can still route them.
-    const runnerDatasource = runner?.state.datasource;
-    const queries: DataQuery[] = (runner?.state.queries ?? []).map((query) =>
-      query.datasource ? query : { ...query, datasource: runnerDatasource }
-    );
-    // Known limitation (follow-up): template variables are sent un-interpolated, so captured
-    // traffic won't match a panel that uses $vars until per-datasource interpolation is applied.
-    if (queries.filter((query) => !query.hide).length === 0) {
-      throw new Error(t('dashboard.diagnostics.no-queries', 'This panel has no active queries to capture.'));
-    }
-    const timeRange = sceneGraph.getTimeRange(panel).state.value;
+  const [{ loading: isGenerating, error }, onDownload] = useAsyncFn(
+    async (includeLogs: boolean) => {
+      const panel = panelRef?.resolve();
+      if (!panel) {
+        return;
+      }
+      const runner = getQueryRunnerFor(panel);
+      // Classic panels keep the datasource on the query runner rather than on each target, and unlike
+      // the normal /api/ds/query path nothing fills that in here. Copy the runner-level datasource
+      // onto any query that lacks one so the diagnostics endpoint can still route them.
+      const runnerDatasource = runner?.state.datasource;
+      const queries: DataQuery[] = (runner?.state.queries ?? []).map((query) =>
+        query.datasource ? query : { ...query, datasource: runnerDatasource }
+      );
+      // Known limitation (follow-up): template variables are sent un-interpolated, so captured
+      // traffic won't match a panel that uses $vars until per-datasource interpolation is applied.
+      if (queries.filter((query) => !query.hide).length === 0) {
+        throw new Error(t('dashboard.diagnostics.no-queries', 'This panel has no active queries to capture.'));
+      }
+      const timeRange = sceneGraph.getTimeRange(panel).state.value;
 
-    const controller = new AbortController();
-    abortRef.current = controller;
-    await downloadDiagnosticsForQueries(queries, String(timeRange.from.valueOf()), String(timeRange.to.valueOf()), {
-      includeLogs,
-      signal: controller.signal,
-    });
-  }, [includeLogs, panelRef]);
+      const controller = new AbortController();
+      abortRef.current = controller;
+      await downloadDiagnosticsForQueries(queries, String(timeRange.from.valueOf()), String(timeRange.to.valueOf()), {
+        includeLogs,
+        signal: controller.signal,
+      });
+    },
+    [panelRef]
+  );
 
   const handleDismiss = () => {
     abortRef.current?.abort();
@@ -111,71 +80,17 @@ function DownloadDiagnosticsRenderer({ model }: SceneComponentProps<DownloadDiag
   };
 
   return (
-    <div>
-      <p className={styles.info}>
+    <DiagnosticsDrawerContent
+      description={
         <Trans i18nKey="dashboard.diagnostics.info-text-panel">
           Generates a diagnostic bundle for this panel by re-running its queries with HTTP capture active. The download
           may take a moment while the bundle is generated.
         </Trans>
-      </p>
-
-      {/* Seed warning. Additional warnings (e.g. sensitive-data redaction, large bundles) are added here later. */}
-      <Alert
-        severity="warning"
-        title={t('dashboard.diagnostics.sensitive-warning-title', 'May contain sensitive data')}
-      >
-        <Trans i18nKey="dashboard.diagnostics.sensitive-warning-body">
-          The bundle can include request headers, query parameters, and server log lines. Review it before sharing
-          outside your organization.
-        </Trans>
-      </Alert>
-
-      <Checkbox
-        id="diagnostics-include-server-logs"
-        label={t('dashboard.diagnostics.include-server-logs', 'Include server logs')}
-        aria-label={t('dashboard.diagnostics.include-server-logs', 'Include server logs')}
-        description={t(
-          'dashboard.diagnostics.include-server-logs-description',
-          'Adds filtered query logs and a bounded unfiltered server-log window to the bundle.'
-        )}
-        checked={includeLogs}
-        disabled={isGenerating}
-        onChange={(event) => setIncludeLogs(event.currentTarget.checked)}
-      />
-
-      {error && (
-        <Alert severity="error" title={t('dashboard.diagnostics.error-title', 'Failed to generate diagnostics')}>
-          {diagnosticsErrorMessage(error)}
-        </Alert>
-      )}
-
-      <div
-        className={styles.buttonRow}
-        role="group"
-        aria-label={t('dashboard.diagnostics.actions', 'Diagnostics actions')}
-      >
-        <Button variant="primary" onClick={onDownload} disabled={isGenerating} icon="download-alt">
-          {isGenerating ? (
-            <Trans i18nKey="dashboard.diagnostics.generating-button">Generating…</Trans>
-          ) : (
-            <Trans i18nKey="dashboard.diagnostics.download-button">Download diagnostics</Trans>
-          )}
-        </Button>
-        <Button variant="secondary" onClick={handleDismiss} fill="outline">
-          <Trans i18nKey="dashboard.diagnostics.cancel-button">Cancel</Trans>
-        </Button>
-      </div>
-    </div>
+      }
+      error={error}
+      isGenerating={isGenerating}
+      onDownload={onDownload}
+      onDismiss={handleDismiss}
+    />
   );
 }
-
-const getStyles = (theme: GrafanaTheme2) => ({
-  info: css({
-    marginBottom: theme.spacing(2),
-  }),
-  buttonRow: css({
-    display: 'flex',
-    gap: theme.spacing(2),
-    marginTop: theme.spacing(2),
-  }),
-});
